It is now possible to view queries #521

+++ priv/www/js/tools.js
@@ -320,6 +320,26 @@ function initQueryBuilderDialog(queryDoctype) {
   return dialog;
 }
 
+function initQueryPreviewButton() {
+  $('#view-query-button').button({
+    icons: {primary: "ui-icon-circle-arrow-s"}
+  }).click(function() {
+    $('#query-view').slideToggle();
+  });
+  
+  return false;
+}
+
+function initQueryChooseButton() {
+  $('#choose-query-button').button({
+    icons: {primary: "ui-icon-arrowreturnthick-1-s"}
+  }).click(function() {
+    $('#query-index-listing').slideToggle();
+  });
+  
+  return false;
+}
+
 function initQueryNewButton() {
   $('#new-query-button').button({
     icons: {primary: "ui-icon-plus"}
@@ -432,28 +452,42 @@ function initQuerySaveButton(button, buttonData) {
   button.button({
     icons: {primary: "ui-icon-document"}
   }).click(function (e) {
-    var completeFunction = function() {
-      getQueryEdit(buttonData.attr('data-query-id'));
-      flashHighlight("Success", "Your query has been saved.");
-    };
+    var bData = buttonData();
     
-    saveQuery(buttonData, completeFunction);
+    if (!bData.length < 1) {
+      var completeFunction = function() {
+        getQueryEdit(bData.attr('data-query-id'));
+        flashHighlight("Success", "Your query has been saved.");
+      };
+      
+      saveQuery(bData, completeFunction);
+    } else {
+      flashHighlight("Info", "No query has been chosen to save.");
+    }
   });
 }
 
 function initQueryDeleteButton(button, buttonData) {
-  button.button({icons: {primary: "ui-icon-trash"}}).click(function (e) {
-    var deleteButton = $(e.target);
-    var queryId = buttonData.attr('data-query-id');
-    var queryRev = buttonData.attr('data-query-rev');
-    var completeMessage = "Your query has been deleted.";
-    var completeFunction = function() {
-      button.parent('div').parent('div').empty();
-      initQueryIndex();
-    };
+  button.button(
+    {icons: {primary: "ui-icon-trash"}
+  }).click(function (e) {
+    var bData = buttonData();
     
-    if (confirm("Are you sure?")) {
-      deleteQuery(queryId, queryRev, completeMessage, completeFunction);
+    if (!bData.length < 1) {
+      var deleteButton = $(e.target);
+      var queryId = bData.attr('data-query-id');
+      var queryRev = bData.attr('data-query-rev');
+      var completeMessage = "Your query has been deleted.";
+      var completeFunction = function() {
+        button.parent('div').parent('div').empty();
+        initQueryIndex();
+      };
+      
+      if (confirm("Are you sure?")) {
+        deleteQuery(queryId, queryRev, completeMessage, completeFunction);
+      }
+    } else {
+      flashHighlight("Info", "No query has been chosen to delete.");
     }
   });
   
@@ -464,7 +498,13 @@ function initQueryAddConditionButton(button, buttonData) {
   button.button({
     icons: {primary: "ui-icon-plus"}
   }).click(function (e) {
-    initQueryBuilderDialog(buttonData.attr('data-query-doctype')).dialog("open");
+    var bData = buttonData();
+    
+    if (!bData.length < 1) {
+      initQueryBuilderDialog(bData.attr('data-query-doctype')).dialog("open");
+    } else {
+      flashHighlight("Info", "You must choose a query first.");
+    }
   });
 
   return false;
@@ -476,11 +516,11 @@ function getQueryEdit(queryId) {
   
   $.get(url, function(queryData) {
     target.html(queryData);
-    initQueryEditButtons($('#query-editing-data'));
     // TODO don't repeat this code. It is also in initQueryBuilderDialog
     tableBody = $('#query-conditions-listing tbody');
     tableBody.sortable();
     initConditionRemoveButtons(tableBody);
+    getQueryView();
   });
   
   return false;
@@ -494,14 +534,93 @@ function initQueryIndex() {
     target.html(index);
     target.click(function(e) {
       getQueryEdit($(e.target).attr('data-query-id'));
+      target.slideToggle();
     });
   });
+}
+
+function getQueryView(startkey, startid, prevkeys, previds) {
+  var queryInfo = $('#query-editing-data');
+  var queryId = queryInfo.attr('data-query-id');
+  var url = "queries/" + queryId + "/view?";
+  var limit = $('#query-limit').val() * 1;
+  
+  // Initialize some values if we're at the beginning of the listing
+  if (!prevkeys) {
+    startkey = $('#query-filter').val();
+    prevkeys = [];
+    previds = [];
+  }
+  
+  if (startkey) {
+    url = url + '&startkey=' + JSON.stringify([startkey]);
+    
+    if (startid) {
+      url = url + '&startkey_id=' + startid;
+    }
+  }
+  
+  // The user supplied limit will need a plus one so that we can
+  // get the start key for the next page from the server.
+  if (limit) {
+    url = url + '&limit=' + (limit + 1);
+  } else {
+    // Ten is the default and I don't let people leave it blank
+    // because the list could be huge.
+    $('#query-limit').val(10);
+    url = url + '&limit=11';
+  }
+  
+  $.get(url, function(data) {
+    $('#query-list-view').html(data);
+    
+    $('#previous-page').button({
+      icons: {primary:'ui-icon-circle-arrow-w'} 
+    }).click(function() {
+      getQueryView(prevkeys.pop(), previds.pop(), prevkeys, previds);
+    });
+    
+    // Collect the values needed for paging from the HTML
+    $('#next-page').button({
+      icons: {secondary:'ui-icon-circle-arrow-e'}
+    }).click(function() {
+      var nextkey = $(this).attr('data-startkey');
+      var nextid = $(this).attr('data-startid');
+      var prevkey = $('#first-index-element').attr('data-first-key');
+      var previd = $('#first-index-element').attr('data-first-id');
+      prevkeys.push(prevkey);
+      previds.push(previd);
+      
+      getQueryView(nextkey, nextid, prevkeys, previds);
+    });
+    
+    // Disable the previous button if we're at the beginning
+    if (prevkeys.length == 0) {
+      $('#previous-page').button("disable");
+    }
+    
+    // Disable the next button if we're at the end
+    if ($('#next-page').attr('data-last-page')) {
+      $('#next-page').button("disable");
+    }
+  
+    $('nav.pager').buttonset();
+    
+  });
+
 }
 
 $(function () {
   initTabs(); 
   $('#query-builder-dialog').hide();
   $('#query-new-dialog').hide();
+  initQueryEditButtons(function () {return $('#query-editing-data')});
   initQueryNewButton();
+  initQueryChooseButton();
+  initQueryPreviewButton();
+  $('#button-bar').buttonset();
   initQueryIndex();
+  $('#query-filter-form input').keyup(function() {
+    getQueryView();
+  });
 });
